+++ src/Vivid/C.php
@@ -60,7 +60,7 @@ use Charm\Vivid\Kernel\Interfaces\ModuleInterface;
 class C
 {
     /** @var string the version of charm */
-    public const VERSION = "3.6";
+    public const string VERSION = "3.7";
 
     /**
      * Get a loaded module