+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Redis usage in the whole system is now optional, works fine without redis at all [#40]
 - The environment config file `app/app.env` is now a proper ini-file. 
   Kept legacy support but preferred syntax is now e.g. `ENVIRONMENT=Prod`.
+- The cache clear command is now available via `bob c:clr` to make calling it more efficient.
 
 ---