+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Database migrations are now only run on app-related modules, no kernel modules, since they have no migrations.
 - The keys for the authentication token on requests can now be adjusted via config.
 - Improve backup + restore methods for models. Now with an optimized handling of large datasets and using NDJSON.
+- Updated dependency packages to their latest versions, replace maximebf/debugbar with php-debugbar.
 
 ### 🔥 Removed